Extract TokensMixin + CompletionsMixin + HoverMixin from A816LanguageServer
