+++ a816/lsp/server.py
@@ -4,7 +4,7 @@ import os
 import re
 from collections.abc import Callable
 from pathlib import Path
-from typing import Any, ClassVar
+from typing import Any
 
 try:
     import tomllib
@@ -12,8 +12,6 @@ except ModuleNotFoundError:  # pragma: no cover
     tomllib = None  # type: ignore[assignment]
 
 from lsprotocol.types import (
-    CompletionItem,
-    CompletionItemKind,
     CompletionList,
     CompletionParams,
     Diagnostic,
@@ -28,8 +26,6 @@ from lsprotocol.types import (
     Hover,
     HoverParams,
     Location,
-    MarkupContent,
-    MarkupKind,
     MessageType,
     Position,
     PrepareRenameParams,
@@ -61,53 +57,11 @@ from a816.cpu.cpu_65c816 import AddressingMode, snes_opcode_table
 from a816.exceptions import FormattingError
 from a816.formatter import A816Formatter, FormattingOptions
 from a816.lsp.document import A816Document
+from a816.lsp.handlers.completions import CompletionsMixin
+from a816.lsp.handlers.hover import HoverMixin
+from a816.lsp.handlers.tokens import TokensMixin
 from a816.lsp.mask import build_code_mask
 from a816.lsp.workspace import WorkspaceIndex
-from a816.parse.ast.nodes import (
-    AllocAstNode,
-    AsciiAstNode,
-    AssignAstNode,
-    AstNode,
-    BinOp,
-    CastAccessExprNode,
-    CastValueExprNode,
-    CodeLookupAstNode,
-    CodePositionAstNode,
-    CodeRelocationAstNode,
-    CommentAstNode,
-    DataNode,
-    DebugAstNode,
-    DocstringAstNode,
-    ExpressionAstNode,
-    ExprNode,
-    ExternAstNode,
-    ForAstNode,
-    IfAstNode,
-    ImportAstNode,
-    IncludeAstNode,
-    IncludeBinaryAstNode,
-    IncludeIpsAstNode,
-    LabelAstNode,
-    LabelDeclAstNode,
-    MacroApplyAstNode,
-    MacroAstNode,
-    MapAstNode,
-    OpcodeAstNode,
-    Parenthesis,
-    PoolAstNode,
-    ReclaimAstNode,
-    RegisterSizeAstNode,
-    RelocateAstNode,
-    ScopeAstNode,
-    StructAstNode,
-    SymbolAffectationAstNode,
-    TableAstNode,
-    Term,
-    TextAstNode,
-    UnaryOp,
-)
-from a816.parse.scanner_states import KEYWORDS
-from a816.parse.tokens import TokenType
 from a816.util import uri_to_path
 
 logger = logging.getLogger(__name__)
@@ -117,7 +71,7 @@ FILE_URI_PREFIX = "file://"
 __all__ = ["A816Document", "A816LanguageServer", "WorkspaceIndex", "lsp_main"]
 
 
-class A816LanguageServer:
+class A816LanguageServer(CompletionsMixin, HoverMixin, TokensMixin):
     """Enhanced LSP server for a816 assembly language"""
 
     def __init__(self) -> None:
@@ -281,327 +235,6 @@ class A816LanguageServer:
         if workspace:
             await asyncio.to_thread(workspace.replace_document, doc)
         self._publish_diagnostics(params.text_document.uri, doc.diagnostics)
-
-    def _local_completions(self, doc: A816Document) -> list[CompletionItem]:
-        items: list[CompletionItem] = []
-        for label in doc.labels:
-            items.append(
-                CompletionItem(
-                    label=label,
-                    kind=CompletionItemKind.Function,
-                    detail="Label",
-                    documentation=doc.label_docstrings.get(label.lower()),
-                )
-            )
-        for symbol in doc.symbols:
-            items.append(CompletionItem(label=symbol, kind=CompletionItemKind.Variable, detail="Symbol"))
-        for macro_name in doc.macros:
-            macro_parameters = doc.macro_params.get(macro_name, [])
-            param_sig = f"({', '.join(macro_parameters)})" if macro_parameters else "()"
-            macro_doc = doc.macro_docstrings.get(macro_name.lower())
-            documentation = macro_doc or (
-                f"User-defined macro with {len(macro_parameters)} parameters"
-                if macro_parameters
-                else "User-defined macro"
-            )
-            items.append(
-                CompletionItem(
-                    label=macro_name,
-                    kind=CompletionItemKind.Function,
-                    detail=f"User Macro{param_sig}",
-                    documentation=documentation,
-                )
-            )
-        return items
-
-    def _handle_completions(self, params: CompletionParams) -> CompletionList:
-        doc = self.documents.get(params.text_document.uri)
-        if not doc:
-            return CompletionList(is_incomplete=False, items=[])
-        line_num = params.position.line
-        if line_num >= len(doc.lines):
-            return CompletionList(is_incomplete=False, items=[])
-
-        line = doc.lines[line_num]
-        char_pos = params.position.character
-        word_start = char_pos
-        while word_start > 0 and line[word_start - 1].isalnum():
-            word_start -= 1
-        current_word = line[word_start:char_pos].lower()
-
-        # Context-aware: cursor after `in` / `into` / `.reclaim` → pool names only.
-        pool_items = self._pool_name_completions_in_context(line, char_pos, doc)
-        if pool_items is not None:
-            filtered = (
-                [item for item in pool_items if item.label.lower().startswith(current_word)]
-                if current_word
-                else pool_items
-            )
-            return CompletionList(is_incomplete=False, items=filtered[:50])
-
-        all_items: list[CompletionItem] = []
-        all_items.extend(self._opcode_completions)
-        all_items.extend(self._keyword_completions)
-        all_items.extend(self._register_completions)
-        all_items.extend(self._build_labels_completions(doc))
-        all_items.extend(self._local_completions(doc))
-        workspace = self._ensure_workspace_index()
-        if workspace:
-            all_items.extend(self._build_workspace_label_completions(doc, workspace))
-            all_items.extend(self._build_workspace_symbol_completions(doc, workspace))
-            all_items.extend(self._build_workspace_macro_completions(doc, workspace))
-
-        filtered = (
-            [item for item in all_items if item.label.lower().startswith(current_word)] if current_word else all_items
-        )
-        return CompletionList(is_incomplete=False, items=filtered[:50])
-
-    def _pool_name_completions_in_context(
-        self, line: str, char_pos: int, doc: A816Document
-    ) -> list[CompletionItem] | None:
-        """When the cursor sits where a pool name is expected (after
-        `.alloc X in `, `.relocate X N N into `, or `.reclaim `), suggest
-        only declared pool names. Returns None when context doesn't match
-        so the default completion list is used."""
-        prefix = line[:char_pos].rstrip()
-        triggers = (".alloc", ".relocate", ".reclaim")
-        if not any(t in prefix for t in triggers):
-            return None
-        # Heuristic: cursor follows the `in` / `into` keyword (alloc / relocate)
-        # or `.reclaim NAME` (after pool name comes addresses, but the pool name
-        # is the first token after `.reclaim`).
-        # Prefix contains at least one trigger token so split() is non-empty.
-        last = prefix.split()[-1].lower()
-        if last not in ("in", "into", ".reclaim"):
-            return None
-        workspace = self._ensure_workspace_index()
-        names: set[str] = set(doc.pools)
-        if workspace:
-            names |= set(workspace.pools)
-        return [CompletionItem(label=name, kind=CompletionItemKind.Module, detail=".pool") for name in sorted(names)]
-
-    @staticmethod
-    def _word_span(line: str, char_pos: int) -> tuple[int, int]:
-        start = char_pos
-        while start > 0 and (line[start - 1].isalnum() or line[start - 1] == "_"):
-            start -= 1
-        end = char_pos
-        while end < len(line) and (line[end].isalnum() or line[end] == "_"):
-            end += 1
-        return start, end
-
-    @staticmethod
-    def _markdown_hover(value: str) -> Hover:
-        return Hover(contents=MarkupContent(kind=MarkupKind.Markdown, value=value))
-
-    def _hover_for_opcode_or_keyword(self, base_word: str, word: str) -> Hover | None:
-        if base_word in snes_opcode_table:
-            return self._markdown_hover(
-                f"**{word.upper()}** - 65c816 Instruction\n\nSupported addressing modes: {len(snes_opcode_table[base_word])}"
-            )
-        if base_word in KEYWORDS:
-            return self._markdown_hover(f"**{word.upper()}** - Assembler directive")
-        return None
-
-    def _hover_for_label_or_scope(
-        self, doc: A816Document, workspace: WorkspaceIndex | None, raw_word: str, word: str
-    ) -> Hover | None:
-        label_doc = doc.label_docstrings.get(word) or (workspace.get_label_doc(raw_word) if workspace else None)
-        if label_doc:
-            return self._markdown_hover(f"**{raw_word}**\n\n{label_doc}")
-        scope_doc = doc.scope_docstrings.get(word) or (workspace.get_scope_doc(raw_word) if workspace else None)
-        if scope_doc:
-            return self._markdown_hover(f"**{raw_word}**\n\n{scope_doc}")
-        return None
-
-    def _hover_for_macro(
-        self, doc: A816Document, workspace: WorkspaceIndex | None, raw_word: str, word: str
-    ) -> Hover | None:
-        macro_name = next((name for name in doc.macros if name.lower() == word), None)
-        if macro_name:
-            doc_text = doc.macro_docstrings.get(word)
-            if doc_text:
-                params_str = self._format_macro_params(doc.macro_params.get(macro_name, []))
-                return self._markdown_hover(f"**{macro_name}{params_str}**\n\n{doc_text}")
-        if workspace:
-            macro_location = workspace.get_macro_location(raw_word)
-            if macro_location:
-                _, actual = macro_location
-                doc_text = workspace.get_macro_doc(actual)
-                if doc_text:
-                    params_str = self._format_macro_params(workspace.get_macro_params(actual))
-                    return self._markdown_hover(f"**{actual}{params_str}**\n\n{doc_text}")
-        return None
-
-    @staticmethod
-    def _format_macro_params(param_list: list[str]) -> str:
-        return f"({', '.join(param_list)})" if param_list else "()"
-
-    def _handle_hover(self, params: HoverParams) -> Hover | None:
-        doc = self.documents.get(params.text_document.uri)
-        if not doc:
-            return None
-        line_num = params.position.line
-        if line_num >= len(doc.lines):
-            return None
-        line = doc.lines[line_num]
-
-        workspace = self._ensure_workspace_index()
-        directive_hover = self._hover_for_module_directive(line, params.text_document.uri, workspace)
-        if directive_hover:
-            return directive_hover
-
-        word_start, word_end = self._word_span(line, params.position.character)
-        if word_start >= word_end:
-            return None
-        raw_word = line[word_start:word_end]
-        word = raw_word.lower()
-        base_word = word.split(".")[0] if "." in word else word
-
-        opcode_or_keyword = self._hover_for_opcode_or_keyword(base_word, word)
-        if opcode_or_keyword:
-            return opcode_or_keyword
-        pool_hover = self._hover_for_pool_directive(doc, raw_word)
-        if pool_hover:
-            return pool_hover
-        dotted_word = self._dotted_word_span(line, params.position.character)
-        struct_hover = self._hover_for_struct_field(dotted_word)
-        if struct_hover:
-            return struct_hover
-        return self._hover_for_label_or_scope(doc, workspace, raw_word, word) or self._hover_for_macro(
-            doc, workspace, raw_word, word
-        )
-
-    @staticmethod
-    def _dotted_word_span(line: str, char_pos: int) -> str:
-        """Like `_word_span` but stretches across `.` boundaries so
-        `Type.field.mask` extracts as one string."""
-        start = char_pos
-        while start > 0 and (line[start - 1].isalnum() or line[start - 1] in "_."):
-            start -= 1
-        end = char_pos
-        while end < len(line) and (line[end].isalnum() or line[end] in "_."):
-            end += 1
-        return line[start:end]
-
-    def _hover_for_struct_field(self, word: str) -> Hover | None:
-        """Auto-doc for `Type.field`, `Type.field.mask`, `Type.field.shift`.
-
-        Reaches into the workspace-shared struct registry built during AST
-        indexing. Bit-field aux symbols get computed mask / shift values
-        inlined into the hover so the user doesn't need to mentally compute
-        them while editing.
-        """
-        struct, field, aux = self._parse_struct_field_path(word)
-        if struct is None or field is None:
-            return None
-        meta = self._struct_field_meta(struct, field)
-        if meta is None:
-            return None
-        field_type, bit_width, mask, shift = meta
-        if aux == "mask":
-            body = (
-                f"**`{struct}.{field}.mask`** = `0x{mask:02X}`\n\n"
-                f"Pre-shifted bit-mask for the `{field}` field "
-                f"({bit_width}-bit, shift {shift}). Use as an immediate operand."
-            )
-            return self._markdown_hover(body)
-        if aux == "shift":
-            body = (
-                f"**`{struct}.{field}.shift`** = `{shift}`\n\n"
-                f"LSB position of the `{field}` field inside its containing byte."
-            )
-            return self._markdown_hover(body)
-        if bit_width is not None:
-            body = (
-                f"**`{struct}.{field}`** — `{field_type}` bit-field\n\n"
-                f"Width: `{bit_width}` bits, shift `{shift}`, mask `0x{mask:02X}`."
-            )
-        else:
-            body = f"**`{struct}.{field}`** — `{field_type}` field of struct `{struct}`."
-        return self._markdown_hover(body)
-
-    @staticmethod
-    def _parse_struct_field_path(word: str) -> tuple[str | None, str | None, str | None]:
-        """Split `Type.field` / `Type.field.mask` / `Type.field.shift` into parts."""
-        parts = word.split(".")
-        if len(parts) == 2:
-            return parts[0], parts[1], None
-        if len(parts) == 3 and parts[2] in ("mask", "shift"):
-            return parts[0], parts[1], parts[2]
-        return None, None, None
-
-    _BIT_FIELD_TYPE_RE: ClassVar[re.Pattern[str]] = re.compile(r"u(\d+)")
-
-    def _struct_field_meta(self, struct_name: str, field_name: str) -> tuple[str, int | None, int, int] | None:
-        """Return `(declared_type, bit_width, mask, shift)` for a struct field, or None."""
-        for doc in self.documents.values():
-            for node in doc.ast_nodes:
-                if isinstance(node, StructAstNode) and node.name == struct_name:
-                    meta = self._field_meta_in_struct(node, field_name)
-                    if meta is not None:
-                        return meta
-        return None
-
-    def _field_meta_in_struct(self, node: StructAstNode, field_name: str) -> tuple[str, int | None, int, int] | None:
-        bit_position = 0
-        for fname, ftype in node.fields:
-            bit_match = self._BIT_FIELD_TYPE_RE.fullmatch(ftype)
-            if bit_match:
-                width = int(bit_match.group(1))
-                if fname == field_name:
-                    shift = bit_position % 8
-                    mask = ((1 << width) - 1) << shift
-                    return ftype, width, mask, shift
-                bit_position += width
-                continue
-            bit_position = 0  # primitive flushes the current bit run
-            if fname == field_name:
-                return ftype, None, 0, 0
-        return None
-
-    def _hover_for_pool_directive(self, doc: A816Document, word: str) -> Hover | None:
-        if word in doc.pools:
-            detail = doc.pool_details.get(word, "")
-            consumer_count = len(doc.pool_consumers.get(word, []))
-            body = f"**`.pool {word}`**\n\n{detail}\n\n{consumer_count} consumer(s) in this document"
-            return self._markdown_hover(body)
-        if word in doc.allocs:
-            pool_name = doc.alloc_target_pool.get(word, "?")
-            detail = doc.pool_details.get(pool_name, "")
-            kind = ".relocate" if word in doc.allocs and pool_name in doc.alloc_target_pool.values() else ".alloc"
-            del kind  # cannot disambiguate cheaply; show generic
-            body = f"**`{word}`** in pool `{pool_name}`\n\n{detail}"
-            return self._markdown_hover(body)
-        return None
-
-    def _hover_for_module_directive(
-        self, line: str, current_uri: str, workspace: WorkspaceIndex | None
-    ) -> Hover | None:
-        """Return the target module's leading docstring when cursor lands on
-        an `.include "path"` or `.import "module"` line."""
-        if workspace is None:
-            return None
-        include_match = re.search(r'\.include\s+[\'"]([^\'"]+)[\'"]', line, re.IGNORECASE)
-        import_match = re.search(r'\.import\s+[\'"]([^\'"]+)[\'"]', line, re.IGNORECASE)
-        target_uri: str | None = None
-        label: str | None = None
-        if include_match:
-            resolved = self._resolve_include_path(include_match.group(1), current_uri)
-            if resolved:
-                target_uri = Path(resolved).as_uri()
-                label = include_match.group(1)
-        elif import_match:
-            resolved = self._resolve_module_path(import_match.group(1), current_uri)
-            if resolved:
-                target_uri = Path(resolved).as_uri()
-                label = import_match.group(1)
-        if not target_uri or label is None:
-            return None
-        docstring = workspace.module_docstrings.get(target_uri)
-        if not docstring:
-            return None
-        return self._markdown_hover(f"**{label}**\n\n{docstring}")
 
     @staticmethod
     def _doc_symbol_range(pos: Position, name_len: int) -> Range:
@@ -976,12 +609,6 @@ class A816LanguageServer:
         )
         return SignatureHelp(signatures=[signature_info], active_signature=0, active_parameter=0)
 
-    def _handle_semantic_tokens_full(self, params: SemanticTokensParams) -> SemanticTokens | None:
-        doc = self.documents.get(params.text_document.uri)
-        if not doc:
-            return None
-        return SemanticTokens(data=self._analyze_semantic_tokens(doc))
-
     def _handle_format_document(self, params: DocumentFormattingParams) -> list[TextEdit] | None:
         doc = self.documents.get(params.text_document.uri)
         if not doc:
@@ -1044,505 +671,6 @@ class A816LanguageServer:
             start=Position(line=0, character=0),
             end=Position(line=last_line_index, character=end_character),
         )
-
-    def _analyze_semantic_tokens(self, doc: A816Document) -> list[int]:
-        """Analyze document and generate semantic tokens using the actual parser"""
-        tokens: list[int] = []
-        prev_line = 0
-        prev_char = 0
-
-        # Use the parsed AST nodes to generate semantic tokens
-        semantic_tokens = self._extract_semantic_tokens_from_ast(doc)
-
-        # Convert to LSP format (delta encoding)
-        for token in sorted(semantic_tokens, key=lambda t: (t["line"], t["char"])):
-            current_line = token["line"]
-            current_char = token["char"]
-
-            if current_line == prev_line:
-                delta_line = 0
-                delta_char = current_char - prev_char
-            else:
-                delta_line = current_line - prev_line
-                delta_char = current_char
-
-            tokens.extend(
-                [
-                    delta_line,  # deltaLine
-                    delta_char,  # deltaStart
-                    token["length"],  # length
-                    token["type"],  # tokenType
-                    0,  # tokenModifiers
-                ]
-            )
-
-            # Update previous position
-            prev_line = current_line
-            prev_char = current_char
-
-        return tokens
-
-    def _extract_semantic_tokens_from_ast(self, doc: A816Document) -> list[dict[str, Any]]:
-        """Extract semantic tokens from parsed AST nodes, with a line-based
-        fallback when parsing failed hard (scanner errors leave `ast_nodes`
-        empty, which would otherwise produce zero highlights and a plain-text
-        document in the editor)."""
-        tokens: list[dict[str, Any]] = []
-
-        logger.debug(f"Processing {len(doc.ast_nodes)} AST nodes for semantic tokens")
-
-        if not doc.ast_nodes:
-            if doc.parse_error:
-                logger.info("Falling back to line tokenizer: %s", doc.parse_error.message)
-            return self._line_based_tokens(doc)
-
-        for node in doc.ast_nodes:
-            self._visit_node_for_tokens(node, tokens, doc)
-
-        logger.debug(f"Generated {len(tokens)} AST-only tokens")
-        return tokens
-
-    def _line_based_tokens(self, doc: A816Document) -> list[dict[str, Any]]:
-        """Best-effort highlighting when the AST is unavailable."""
-        tokens: list[dict[str, Any]] = []
-        for idx, line in enumerate(doc.lines):
-            tokens.extend(self._tokenize_line(line, idx))
-        return tokens
-
-    _DIRECTIVE_TYPES: ClassVar[tuple[type, ...]] = (
-        MacroApplyAstNode,
-        CodePositionAstNode,
-        CodeRelocationAstNode,
-        MapAstNode,
-        IfAstNode,
-        ForAstNode,
-        MacroAstNode,
-        AssignAstNode,
-        SymbolAffectationAstNode,
-        ExternAstNode,
-        ImportAstNode,
-        StructAstNode,
-        DataNode,
-        AsciiAstNode,
-        TextAstNode,
-        TableAstNode,
-        IncludeBinaryAstNode,
-        IncludeIpsAstNode,
-        ScopeAstNode,
-        PoolAstNode,
-        AllocAstNode,
-        RelocateAstNode,
-        ReclaimAstNode,
-        DebugAstNode,
-        LabelDeclAstNode,
-        RegisterSizeAstNode,
-        CodeLookupAstNode,
-    )
-
-    @staticmethod
-    def _semantic_token_type(node: AstNode) -> int | None:
-        if isinstance(node, LabelAstNode):
-            return 1  # function (label)
-        if isinstance(node, OpcodeAstNode):
-            return 0  # keyword (opcode)
-        if isinstance(node, CommentAstNode):
-            return 2  # comment
-        if isinstance(node, DocstringAstNode):
-            return 4  # string
-        if isinstance(node, IncludeAstNode) or isinstance(node, A816LanguageServer._DIRECTIVE_TYPES):
-            return 7  # macro (directive)
-        return None
-
-    @staticmethod
-    def _terminates_after_emit(node: AstNode) -> bool:
-        return isinstance(node, DocstringAstNode | IncludeAstNode)
-
-    def _visit_token_children(self, node: AstNode, tokens: list[dict[str, Any]], doc: A816Document) -> None:
-        for attr in ("body", "block", "else_block", "value", "expression", "min_value", "max_value"):
-            child = getattr(node, attr, None)
-            if isinstance(child, list):
-                for entry in child:
-                    if isinstance(entry, AstNode):
-                        self._visit_node_for_tokens(entry, tokens, doc)
-            elif isinstance(child, AstNode):
-                self._visit_node_for_tokens(child, tokens, doc)
-
-    def _visit_node_for_tokens(self, node: AstNode, tokens: list[dict[str, Any]], doc: A816Document) -> None:
-        """Recursively visit AST nodes to extract semantic tokens."""
-        try:
-            if not node.file_info or not node.file_info.position:
-                return
-            if node.file_info.position.file.filename and node.file_info.position.file.filename != doc.uri:
-                return
-
-            pos = node.file_info.position
-            token_text = node.file_info.value
-            type_id = self._semantic_token_type(node)
-            if type_id is not None:
-                tokens.append({"line": pos.line, "char": pos.column, "length": len(token_text), "type": type_id})
-
-            if isinstance(node, OpcodeAstNode) and node.operand:
-                self._visit_node_for_tokens(node.operand, tokens, doc)
-                return
-            if isinstance(node, ExpressionAstNode):
-                self._analyze_expression_tokens(node, tokens)
-                return
-            if self._terminates_after_emit(node):
-                return
-
-            self._visit_token_children(node, tokens, doc)
-        except (AttributeError, KeyError, IndexError, TypeError) as e:
-            logger.debug(f"Error processing AST node {type(node).__name__}: {e}")
-
-    def _analyze_expression_tokens(self, expr_node: ExpressionAstNode, tokens: list[dict[str, Any]]) -> None:
-        """Highlight every token inside an expression — numbers, identifiers,
-        operators, parens, and the `as TYPE` cast wrapping.
-        """
-        try:
-            for expr_part in expr_node.tokens:
-                self._highlight_expr_part(expr_part, tokens)
-        except (AttributeError, KeyError, IndexError, TypeError) as e:
-            logger.debug(f"Error analyzing expression tokens: {e}")
-
-    def _highlight_expr_part(self, part: ExprNode, tokens: list[dict[str, Any]]) -> None:
-        if isinstance(part, CastValueExprNode | CastAccessExprNode):
-            # Recurse into the cast's inner expression — same shape as the
-            # outer ExpressionAstNode token list.
-            for inner in part.inner:
-                self._highlight_expr_part(inner, tokens)
-            return
-        if isinstance(part, BinOp | UnaryOp):
-            tok = part.token
-            if tok.position:
-                tokens.append(
-                    {
-                        "line": tok.position.line,
-                        "char": tok.position.column,
-                        "length": len(tok.value),
-                        "type": 5,  # operator
-                    }
-                )
-            return
-        if isinstance(part, Parenthesis):
-            return  # parens carry no semantic colour of their own
-        if not isinstance(part, Term):
-            return
-        expr_token = part.token
-        if not expr_token.position:
-            return
-        match expr_token.type:
-            case TokenType.NUMBER:
-                tokens.append(
-                    {
-                        "line": expr_token.position.line,
-                        "char": expr_token.position.column,
-                        "length": len(expr_token.value),
-                        "type": 3,  # number
-                    }
-                )
-            case TokenType.QUOTED_STRING:
-                tokens.append(
-                    {
-                        "line": expr_token.position.line,
-                        "char": expr_token.position.column,
-                        "length": len(expr_token.value),
-                        "type": 4,  # string
-                    }
-                )
-            case TokenType.IDENTIFIER:
-                tokens.append(
-                    {
-                        "line": expr_token.position.line,
-                        "char": expr_token.position.column,
-                        "length": len(expr_token.value),
-                        "type": 6,  # variable
-                    }
-                )
-
-    def _classify_identifier(self, identifier: str, doc: A816Document) -> int:
-        """Classify an identifier as a specific token type"""
-        # Check if it's a known label
-        if identifier in doc.labels:
-            return 1  # function (label)
-
-        # Check if it's a known macro
-        if identifier in doc.macros:
-            return 7  # macro (user-defined macro)
-
-        # Check if it's a known symbol
-        if identifier in doc.symbols:
-            return 6  # variable (symbol)
-
-        # Check if it's a directive or keyword
-        if identifier.lower() in KEYWORDS:
-            return 7  # macro (directive)
-
-        # Check if it's an opcode
-        base_identifier = identifier.lower().split(".")[0]
-        if base_identifier in snes_opcode_table:
-            return 0  # keyword (opcode)
-
-        # Check if it's a macro parameter (within macro scope)
-        # This would require context tracking, but for now treat as variable
-
-        # Default to variable for unknown identifiers
-        return 6  # variable
-
-    def _tokenize_line(self, line: str, line_num: int) -> list[dict[str, Any]]:
-        """Tokenize a single line for semantic highlighting"""
-        tokens: list[dict[str, Any]] = []
-
-        # Handle comments first
-        comment_pos = line.find(";")
-        if comment_pos != -1:
-            tokens.append(
-                {
-                    "line": line_num,
-                    "char": comment_pos,
-                    "length": len(line) - comment_pos,
-                    "type": 2,  # comment
-                }
-            )
-            # Only process text before comment
-            line = line[:comment_pos]
-
-        stripped = line.strip()
-        if not stripped:
-            return tokens
-
-        # Check for label
-        label_match = re.match(r"^(\s*)([a-zA-Z_]\w*):(.*)$", line, flags=re.ASCII)
-        if label_match:
-            indent = len(label_match.group(1))
-            label_name = label_match.group(2)
-            tokens.append(
-                {
-                    "line": line_num,
-                    "char": indent,
-                    "length": len(label_name),
-                    "type": 1,  # function (label)
-                }
-            )
-            # Continue with rest of line
-            rest = label_match.group(3).strip()
-            if rest:
-                rest_pos = line.find(rest, indent + len(label_name) + 1)
-                tokens.extend(self._tokenize_instruction(rest, line_num, rest_pos))
-        else:
-            # Regular instruction line
-            first_non_space = len(line) - len(line.lstrip())
-            tokens.extend(self._tokenize_instruction(stripped, line_num, first_non_space))
-
-        return tokens
-
-    def _tokenize_instruction(self, instruction: str, line_num: int, start_pos: int) -> list[dict[str, Any]]:
-        """Tokenize an instruction for semantic highlighting"""
-        tokens: list[dict[str, Any]] = []
-        parts = instruction.split()
-
-        if not parts:
-            return tokens
-
-        # Find the actual position of the first part (opcode/directive)
-        opcode = parts[0]
-        opcode_pos = instruction.find(opcode)
-        if opcode_pos == -1:
-            opcode_pos = 0
-
-        base_opcode = opcode.lower().split(".")[0]
-
-        if base_opcode in snes_opcode_table:
-            # It's an opcode
-            tokens.append(
-                {
-                    "line": line_num,
-                    "char": start_pos + opcode_pos,
-                    "length": len(opcode),
-                    "type": 0,  # keyword
-                }
-            )
-        elif base_opcode in KEYWORDS:
-            # It's a directive
-            tokens.append(
-                {
-                    "line": line_num,
-                    "char": start_pos + opcode_pos,
-                    "length": len(opcode),
-                    "type": 7,  # macro (directive)
-                }
-            )
-
-        # Handle operands
-        if len(parts) > 1:
-            operand_text = " ".join(parts[1:])
-            operand_pos = instruction.find(operand_text, opcode_pos + len(opcode))
-            if operand_pos != -1:
-                tokens.extend(self._tokenize_operand(operand_text, line_num, start_pos + operand_pos))
-
-        return tokens
-
-    @staticmethod
-    def _consume_hex_digits(operand: str, i: int) -> int:
-        while i < len(operand) and operand[i] in "0123456789ABCDEFabcdef":
-            i += 1
-        return i
-
-    @staticmethod
-    def _consume_decimal(operand: str, i: int) -> int:
-        while i < len(operand) and operand[i].isdigit():
-            i += 1
-        return i
-
-    def _consume_number(self, operand: str, i: int) -> int:
-        """Advance past a number literal (hex / decimal / #-prefixed immediate)."""
-        if operand.startswith(("0x", "0X"), i):
-            return self._consume_hex_digits(operand, i + 2)
-        if operand[i] == "#":
-            i += 1
-            if operand.startswith(("0x", "0X"), i):
-                return self._consume_hex_digits(operand, i + 2)
-            return self._consume_decimal(operand, i)
-        return self._consume_decimal(operand, i)
-
-    @staticmethod
-    def _is_register_at(operand: str, i: int) -> bool:
-        return operand[i].upper() in "XYS" and (i == 0 or not operand[i - 1].isalnum())
-
-    def _tokenize_operand(self, operand: str, line_num: int, start_pos: int) -> list[dict[str, Any]]:
-        """Tokenize operand for semantic highlighting."""
-        tokens: list[dict[str, Any]] = []
-        i = 0
-        while i < len(operand):
-            char = operand[i]
-            if char.isspace():
-                i += 1
-                continue
-            if operand.startswith(("0x", "0X"), i) or char.isdigit() or char == "#":
-                start = i
-                i = self._consume_number(operand, i)
-                tokens.append({"line": line_num, "char": start_pos + start, "length": i - start, "type": 3})
-            elif self._is_register_at(operand, i):
-                tokens.append({"line": line_num, "char": start_pos + i, "length": 1, "type": 6})
-                i += 1
-            elif char in "()[],.+-*&|":
-                tokens.append({"line": line_num, "char": start_pos + i, "length": 1, "type": 5})
-                i += 1
-            else:
-                i += 1
-        return tokens
-
-    def _build_opcode_completions(self) -> list[CompletionItem]:
-        """Build completion items for all opcodes"""
-        completions = []
-        for opcode in snes_opcode_table.keys():
-            completions.append(
-                CompletionItem(
-                    label=opcode.upper(),
-                    kind=CompletionItemKind.Keyword,
-                    detail="65c816 Instruction",
-                    documentation=f"65c816 instruction with {len(snes_opcode_table[opcode])} addressing modes",
-                )
-            )
-
-            # Add size variants
-            for size in ["b", "w", "l"]:
-                completions.append(
-                    CompletionItem(
-                        label=f"{opcode.upper()}.{size.upper()}",
-                        kind=CompletionItemKind.Keyword,
-                        detail=f"65c816 Instruction ({size.upper()})",
-                        documentation=f"65c816 instruction with {size}-size specifier",
-                    )
-                )
-
-        return completions
-
-    def _build_keyword_completions(self) -> list[CompletionItem]:
-        """Build completion items for assembler keywords"""
-        return [
-            CompletionItem(label=keyword.upper(), kind=CompletionItemKind.Keyword, detail="Assembler directive")
-            for keyword in KEYWORDS
-        ]
-
-    def _build_register_completions(self) -> list[CompletionItem]:
-        """Build completion items for registers"""
-        registers = ["X", "Y", "S", "A"]
-        return [
-            CompletionItem(label=reg, kind=CompletionItemKind.Variable, detail="65c816 Register") for reg in registers
-        ]
-
-    def _build_labels_completions(self, doc: A816Document) -> list[CompletionItem]:
-        return [
-            CompletionItem(label=sym, kind=CompletionItemKind.Variable, detail="Symbol") for sym in doc.symbols.keys()
-        ]
-
-    def _build_workspace_label_completions(self, doc: A816Document, workspace: WorkspaceIndex) -> list[CompletionItem]:
-        doc_labels = set(doc.labels.keys())
-        items: list[CompletionItem] = []
-        for label in workspace.labels.keys():
-            if label in doc_labels:
-                continue
-            doc_text = workspace.get_label_doc(label)
-            items.append(
-                CompletionItem(
-                    label=label,
-                    kind=CompletionItemKind.Function,
-                    detail="Label",
-                    documentation=doc_text,
-                )
-            )
-        return items
-
-    def _build_workspace_symbol_completions(self, doc: A816Document, workspace: WorkspaceIndex) -> list[CompletionItem]:
-        doc_symbols = set(doc.symbols.keys())
-        items: list[CompletionItem] = []
-        for symbol in workspace.symbols.keys():
-            if symbol in doc_symbols:
-                continue
-            items.append(CompletionItem(label=symbol, kind=CompletionItemKind.Variable, detail="Symbol"))
-        return items
-
-    def _build_workspace_macro_completions(self, doc: A816Document, workspace: WorkspaceIndex) -> list[CompletionItem]:
-        doc_macros = set(doc.macros.keys())
-        items: list[CompletionItem] = []
-        for macro in workspace.macros.keys():
-            if macro in doc_macros:
-                continue
-            macro_params = workspace.get_macro_params(macro)
-            param_sig = f"({', '.join(macro_params)})" if macro_params else ""
-            macro_doc = workspace.get_macro_doc(macro)
-            documentation = macro_doc or (
-                f"User-defined macro with {len(macro_params)} parameters" if macro_params else "User-defined macro"
-            )
-            items.append(
-                CompletionItem(
-                    label=macro,
-                    kind=CompletionItemKind.Function,
-                    detail=f"User Macro{param_sig}",
-                    documentation=documentation,
-                )
-            )
-        return items
-
-    def _workspace_container_name(self, uri: str, workspace: WorkspaceIndex) -> str | None:
-        try:
-            path = uri_to_path(uri)
-        except ValueError:
-            return None
-        root = workspace.root_path
-        if root:
-            try:
-                rel = path.relative_to(root)
-                parent = rel.parent
-                if parent and parent != Path(""):
-                    return parent.as_posix()
-                return None
-            except ValueError:
-                pass
-        parent = path.parent
-        if parent and parent != Path(""):
-            return parent.as_posix()
-        return None
 
     @staticmethod
     def _quoted_span(line: str, match: re.Match[str]) -> tuple[int, int] | None:
